many to many test

+++ src/database/migrations/20211027211327-posts.js
@@ -4,7 +4,7 @@ module.exports = {
   up: (queryInterface, Sequelize) => {
     return queryInterface.createTable('posts', {
       id: {
-        type: Sequelize.INTEGER.UNSIGNED,
+        type: Sequelize.INTEGER(11).UNSIGNED,
         autoIncrement: true,
         primaryKey: true,
       },
